app/requirements.txt


uvicorn
PyMuPDF
python-dotenv
python-docx
fastapi 
PyMuPDF
fastapi
google-generativeai

